requirements.txt
tensorflow==2.16.2
pillow
ipdb
opencv-python==3.4.13.47
scikit-image
# opendr
cython==0.29.7
scipy==1.2.1
openpyxl==2.6.2
ecos==2.0.7.post1
vtk==9.3.1
traits==5.1.1
cvxpy
fancyimpute==0.3.2
# mayavi==4.7.1
pyopengl==3.1.0
pyqt5==5.15.10
scikit-learn==0.21.3
Bottleneck
flask
torch==1.10.1
torchvision==0.11.2
tqdm